Refactoring to have FlagsConfig property in PipelineContext
