+++ metadata-ingestion/tests/unit/test_pulsar_source.py
@@ -5,9 +5,8 @@ from unittest.mock import patch
 import pytest
 
 from datahub.emitter.mcp import MetadataChangeProposalWrapper
-from datahub.ingestion.api.common import PipelineContext
+from datahub.ingestion.api.common import FlagsConfig, PipelineContext
 from datahub.ingestion.run.pipeline_config import (
-    FlagsConfig,
     PipelineConfig,
     SourceConfig,
 )
